trace width bug when setting autorouter trace width < 12mils.
git-svn-id: http://fritzing.googlecode.com/svn/trunk@5913 85f94c43-f63a-0410-9d22-21439860c375

+++ fritzing/src/autoroute/autorouter.cpp
@@ -61,7 +61,7 @@ void Autorouter::updateRoutingStatus() {
 	m_sketchWidget->updateRoutingStatus(routingStatus, true);
 }
 
-TraceWire * Autorouter::drawOneTrace(QPointF fromPos, QPointF toPos, int width, ViewLayer::ViewLayerSpec viewLayerSpec)
+TraceWire * Autorouter::drawOneTrace(QPointF fromPos, QPointF toPos, double width, ViewLayer::ViewLayerSpec viewLayerSpec)
 {
 	long newID = ItemBase::getNextID();
 	ViewGeometry viewGeometry;
